Cleaned file structure for inactivity package
